refactor: LayoutResourceBaseLocatore generalized
