+++ test/Loader/ServerVarResourceBaseLocatorTest.php
@@ -4,15 +4,14 @@ declare(strict_types=1);
 
 namespace Atoolo\Resource\Test\Loader;
 
-use Atoolo\Resource\Loader\DocumentRootLayoutResourceBaseLocator;
 use Atoolo\Resource\Loader\ResourceLayoutResourceBaseLocator;
-use Atoolo\Resource\Loader\StaticResourceBaseLocator;
+use Atoolo\Resource\Loader\ServerVarResourceBaseLocator;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
 
-#[CoversClass(ResourceLayoutResourceBaseLocator::class)]
-class ResourceLayoutResourceBaseLocatorTest extends TestCase
+#[CoversClass(ServerVarResourceBaseLocator::class)]
+class ServerVarResourceBaseLocatorTest extends TestCase
 {
     /**
      * @var array<string,string>
@@ -29,10 +28,24 @@ class ResourceLayoutResourceBaseLocatorTest extends TestCase
         $_SERVER = $this->saveServerState;
     }
 
-    public function testConstruct(): void
+    public function testWithoutSubDirectory(): void
     {
         $_SERVER['RESOURCE_ROOT'] = 'abc';
-        $locator = new ResourceLayoutResourceBaseLocator();
+        $locator = new ServerVarResourceBaseLocator('RESOURCE_ROOT');
+        $this->assertEquals(
+            'abc',
+            $locator->locate(),
+            'unexpected resource base'
+        );
+    }
+
+    public function testWithSubDirectory(): void
+    {
+        $_SERVER['RESOURCE_ROOT'] = 'abc';
+        $locator = new ServerVarResourceBaseLocator(
+            'RESOURCE_ROOT',
+            'objects'
+        );
         $this->assertEquals(
             'abc/objects',
             $locator->locate(),
@@ -42,7 +55,7 @@ class ResourceLayoutResourceBaseLocatorTest extends TestCase
 
     public function testWithMissingSeverVariable(): void
     {
-        $locator = new ResourceLayoutResourceBaseLocator();
+        $locator = new ServerVarResourceBaseLocator('RESOURCE_ROOT');
 
         $this->expectException(RuntimeException::class);
         $locator->locate();
